Generate layouts from klc files.
.klc files are created with Microsoft Keyboard Layout Creator. Since MKLC
is the usual way to create and distribute keyboard layouts in Windows,
this gives access to a large amount of custom layouts. It also makes it
easier to design your own. It should also help people who have different
scancodes.

The powershell script klc2kblayout.ps1 reads the layout you want to
convert (InputLayout), and translates it to the layout that Windows uses
(OutputLayout). Right now it directly creates C, but it will later use a
custom format when I have a loader created.