Add versioning to the install process
This came about because I had an earlier version of the registry key
installed, and the installer wouldn't create the correct one. That of
course, caused the driver to not load. There is now a new "Version"
value in the registry that the install script can check against. It is
directly on the kblayout key, indicating that I might want to move
"Selected Layout" there.